# Improve performance of some notifier things