Prompted by [a report of a performance regression in notifier performance in v11](https://github.com/realm/realm-cocoa/issues/7316), I wrote a few new benchmarks and optimized some stuff. I haven't yet reproduced the regression the user reported, but I did find a regression in notifier creation perf (which is moderately important as it happens on the main thread) and found some room to optimize things which might be similar to what the user was hitting.

It may be easiest to review this by looking at the commits individually as they're each doing fairly distinct things.

Test | v10.8.1 | master | This PR
 -|-|-|-
Create notifier | 409 us | 12.5 ms | 333 us
Chained object links (depth 0) | 25 us | 22 us | 24 us
Chained object links (depth 1) | 314 ms | 337 ms | 10 ms
Chained object links (depth 2) | 10 ms | 11 ms | 380 us
Chained object links (depth 3) | 375 us | 400 us | 25 us
Chained object lists (depth 0) | 24 us | 22 us | 23 us
Chained object lists (depth 1) | 2.4 s | 3.5 s | 30 ms
Chained object lists (depth 2) | 80 ms | 116 ms | 1 ms
Chained object lists (depth 3) | 2.74 ms |3.9 ms | 64 us

The 100x speedup there is not particularly representative of anything realistic, but it should provide a meaningful speedup even without 99 empty lists in each object.